pyproject.toml
[project]
name = "django-todo-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.6",
    "django-debug-toolbar>=6.0.0",
]
